refactor(albwaf): move onlyAllowedIfBoolEquals to global validators
Relates to STACKITTPR-786
